Ensure we keep our original uid for erroring requests
This ensures we keep the original uid we provided for this call, rather than
passing the value back up the stack.
When we select errors via .Errors().ForUid(x) this is the field we search on,
so we need to preserve the uid the author provided rather than passing back
the uid from another service, otherwise error checking won't find this error
and we may return success.